pyproject.toml
[tool.poetry]
name = "pipelang"
version = "0.1.0"
description = ""
authors = ["Kevin Lu <kevinlu1248@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
openai = "requests"
pytest = "^7.2.2"
tiktoken = "^0.3.0"

[tool.poetry.dev-dependencies]
openai = "^0.27.2"
requests = "^2.28.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
